server/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
crashtest==0.4.1
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
dulwich==0.21.7
fastapi==0.115.7
fastjsonschema==2.21.1
filelock==3.16.1
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jiter==0.8.2
keyring==24.3.1
more-itertools==10.5.0
motor==3.6.1
msgpack==1.1.0
openai==1.60.2
packaging==24.2
pexpect==4.9.0
pkginfo==1.12.0
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.9.2
pyproject_hooks==1.2.0
python-multipart==0.0.20
RapidFuzz==3.10.1
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
shellingham==1.5.4
sniffio==1.3.1
starlette==0.45.3
tomlkit==0.13.2
tqdm==4.67.1
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.28.0
xattr==1.1.0
